+++ libs/textarea/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/textarea.module';
+export * from './lib/textarea/textarea.component';