+++ libs/textarea/src/lib/textarea/textarea.component.ts
@@ -1,0 +1,28 @@
+import {
+  Component,
+  ChangeDetectionStrategy,
+  Input,
+  Output,
+  EventEmitter,
+  ViewEncapsulation,
+} from '@angular/core';
+import { ToJson } from '@spryker/utils';
+
+@Component({
+  selector: 'spy-textarea',
+  templateUrl: './textarea.component.html',
+  styleUrls: ['./textarea.component.less'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  encapsulation: ViewEncapsulation.None,
+})
+export class TextareaComponent {
+  @Input() name = '';
+  @Input() value = '';
+  @Input() disabled = false;
+  @Input() placeholder = '';
+  @Input() rows = 4;
+  @Input() cols = 4;
+  @Input() @ToJson() attrs: Record<string, string> = {};
+  @Input() spyId = '';
+  @Output() valueChange = new EventEmitter<any>();
+}