+++ libs/button/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-next.6](https://github.com/spryker/ui-components/compare/@spryker/button@0.1.0-next.5...@spryker/button@0.1.0-next.6) (2020-02-27)
+
+
+### Bug Fixes
+
+* **alert, form-item, button components:** style changes ([5840c3f](https://github.com/spryker/ui-components/commit/5840c3f9795377c742a332c3b83f9b83b9bf37b2))
+
+
+
+
+
+# [0.1.0-next.5](https://github.com/spryker/ui-components/compare/@spryker/button@0.1.0-next.4...@spryker/button@0.1.0-next.5) (2020-02-27)
+
+**Note:** Version bump only for package @spryker/button
+
+
+
+
+
 # [0.1.0-next.4](https://github.com/spryker/ui-components/compare/@spryker/button@0.1.0-next.3...@spryker/button@0.1.0-next.4) (2020-02-25)